MODTLR-24 move ecsTlrSettings to separate controller
