+++ src/main/java/org/folio/controller/EcsTlrController.java
@@ -8,8 +8,7 @@ import static org.springframework.http.HttpStatus.OK;
 import java.util.UUID;
 
 import org.folio.domain.dto.EcsTlr;
-import org.folio.domain.dto.EcsTlrSettings;
-import org.folio.rest.resource.TlrApi;
+import org.folio.rest.resource.EcsTlrApi;
 import org.folio.service.EcsTlrService;
 import org.folio.service.EcsTlrSettingsService;
 import org.springframework.http.HttpStatus;
@@ -22,7 +21,7 @@ import lombok.extern.log4j.Log4j2;
 @RestController
 @Log4j2
 @AllArgsConstructor
-public class EcsTlrController implements TlrApi {
+public class EcsTlrController implements EcsTlrApi {
 
   private final EcsTlrService ecsTlrService;
   private final EcsTlrSettingsService ecsTlrSettingsService;
@@ -59,25 +58,5 @@ public class EcsTlrController implements TlrApi {
     HttpStatus httpStatus = requestDeleted ? NO_CONTENT : NOT_FOUND;
 
     return ResponseEntity.status(httpStatus).build();
-  }
-
-  @Override
-  public ResponseEntity<EcsTlrSettings> getEcsTlrSettings() {
-    log.debug("getEcsTlrSettings:: ");
-
-    return ecsTlrSettingsService.getEcsTlrSettings()
-      .map(ResponseEntity.status(OK)::body)
-      .orElseGet(() -> ResponseEntity.notFound().build());
-  }
-
-  @Override
-  public ResponseEntity<Void> putEcsTlrSettings(EcsTlrSettings ecsTlrSettings) {
-    log.debug("putEcsTlrSettings:: parameters: {}", () -> ecsTlrSettings);
-
-    return ResponseEntity.status(
-        ecsTlrSettingsService.updateEcsTlrSettings(ecsTlrSettings)
-          .map(entity -> NO_CONTENT)
-          .orElse(NOT_FOUND))
-      .build();
   }
 }
